Fix issue provisionally
https://stackoverflow.com/questions/45217234/issue-with-using-spring-oauth-on-java8-standard-environment

https://cloud.google.com/appengine/docs/standard/java/memcache/
Memcache can be useful for other temporary values. However, when considering whether to store a value solely in the memcache and not backed by other persistent storage, be sure that your application behaves acceptably when the value is suddenly not available. Values can expire from the memcache at any time, and can be expired prior to the expiration deadline set for the value. For example, if the sudden absence of a user's session data would cause the session to malfunction, that data should probably be stored in the datastore in addition to the memcache.